backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Vector database
qdrant-client==1.6.9

# Redis
redis==5.0.1

# Object storage
minio==7.2.0
boto3==1.34.0

# Document processing
grobid-client-python==0.0.7
pdfplumber==0.10.3
PyMuPDF==1.23.8
beautifulsoup4==4.12.2
lxml==4.9.3

# Embeddings
openai==1.3.7
sentence-transformers==2.2.2
scikit-learn==1.4.0

# Dimensionality reduction
umap-learn==0.5.4
tensorflow==2.16.1

# Data processing
pandas==2.1.4
numpy==1.25.2

# Configuration
pydantic==2.5.0
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
